work done

+++ 0x04-more_functions_nested_loops/101-print_number.c
@@ -9,7 +9,7 @@
 void print_number(int n)
 {
 	int i, d, p, f;
-	long j;
+	int j;
 	char num[1000];
 
 	p = 0;
